requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0 